Refactor the AbfsReadFile implementation
